+++ src/cards/vehicles/const.js
@@ -1,3 +1,5 @@
 export const PACKAGE = 'sci-fi-vehicles';
 export const NAME = 'Sci-fi vehicles card';
-export const DESCRIPTION = 'Render sci-fi vehicles management.';
+export const MANUFACTURED = 'Renault';
+export const DESCRIPTION =
+  'Render sci-fi ' + MANUFACTURED + ' vehicles management.';